content/base/test/test_CSP_bug663567.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test if XSLT stylesheet is subject to document's CSP</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<iframe style="width:100%;" id='xsltframe'></iframe>
<iframe style="width:100%;" id='xsltframe2'></iframe>

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

// define the expected output of this test
var header = "this xml file should be formatted using an xsl file(lower iframe should contain xml dump)!";

var index = 0;

function checkAllowed () {
  /*   The policy for this test is:
   *   Content-Security-Policy: default-src 'self'
   *
   *   we load the xsl file using:
   *   <?xml-stylesheet type="text/xsl" href="file_CSP_bug663467_allows.xsl"?>
   */
  try {
    var cspframe = document.getElementById('xsltframe');
    var xsltAllowedHeader = cspframe.contentWindow.document.getElementById('xsltheader').innerHTML;
    is(xsltAllowedHeader, header, "XSLT loaded from 'self' should be allowed!");
  }
  catch (e) {
    ok(false, "Error: could not access content in xsltframe!")
  }
  next();
}

function checkBlocked () {
  /*   The policy for this test is:
   *   Content-Security-Policy: default-src 'self'
   *
   *   we load the xsl file using:
   *   <?xml-stylesheet type="text/xsl"
   *    href="http://example.org/tests/content/base/test/file_CSP_bug663467_blocks.xsl"?>
   */
  try {
    var cspframe = document.getElementById('xsltframe2');
    var xsltBlockedHeader = cspframe.contentWindow.document.getElementById('xsltheader');
    is(xsltBlockedHeader, null, "XSLT loaded from different host should be blocked!");
  }
  catch (e) {
    ok(false, "Error: could not access content in xsltframe2!")
  }
  next();
}

var steps = [
  function() {
    document.getElementById('xsltframe').src = 'file_CSP_bug663567_allows.xml';
    document.getElementById('xsltframe').addEventListener('load', checkAllowed, false);
  },
  function() {
    document.getElementById('xsltframe2').src = 'file_CSP_bug663567_blocks.xml';
    document.getElementById('xsltframe2').addEventListener('load', checkBlocked, false);
  },
  function () {
    SimpleTest.finish();
  }
];

function next() {
  if (index >= steps.length) {
    ok(false, "Shouldn't get here!");
    return;
  }
  try {
    var i = index++;
    steps[i]();
  } catch(ex) {
    ok(false, "Caught exception", ex);
  }
}

SpecialPowers.pushPrefEnv(
  {'set':[["security.csp.speccompliant", true]]},
  function () {
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(next);
  }
);

</script>
</body>
</html>
